feat: add hierarchical report formatter and integrate into main execution flow
